adminWeb/requirements/dev.txt
-r common.txt 

django-debug-toolbar